+++ src/components/calcite-modal/calcite-modal.stories.js
@@ -11,7 +11,9 @@ storiesOf('Modal', module)
       <calcite-modal
         active="${boolean("active", true)}"
         color="${select("color", {blue: "blue", red: "red", none: null}, null)}"
-        size="${select("size", ["small", "medium", "large", "fullscreen"], "small")}"
+        scale="${select("scale", ["s", "m", "l"], "m")}"
+        width="${select("width", ["s", "m", "l"], "s")}"
+        fullscreen="${boolean("fullscreen", false)}"
         docked="${boolean("docked", false)}"
         disable-escape="${boolean("disable-escape", false)}"
         no-padding="${boolean("no-padding", false)}"
@@ -33,7 +35,7 @@ storiesOf('Modal', module)
     return `
       <calcite-modal
         active
-        size="${number("size", 500)}"
+        width="${number("width", 500)}"
       >
         <h3 slot="header">Custom Size</h3>
         <div slot="content">
@@ -54,7 +56,9 @@ storiesOf('Modal', module)
         theme="dark"
         active="${boolean("active", true)}"
         color="${select("color", {blue: "blue", red: "red", none: null}, null)}"
-        size="${select("size", ["small", "medium", "large", "fullscreen"], "small")}"
+        scale="${select("scale", ["s", "m", "l"], "m")}"
+        width="${select("width", ["s", "m", "l"], "s")}"
+        fullscreen="${boolean("fullscreen", false)}"
         docked="${boolean("docked", false)}"
         disable-escape="${boolean("disable-escape", false)}"
         no-padding="${boolean("no-padding", false)}"